rag-knowledge-base-complete_1/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0
langchain==0.0.350
langchain-text-splitters==0.0.1

# Vector database and embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.4

# LLM integration
openai==1.3.5

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
pathlib2==2.3.7.post1

# Optional: For better PDF processing
pdfplumber==0.10.3

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2